Improve SelectField's handling of option lists
The class's constructor now checks for a parsing mode to determine how
to initialise its options list. The previous behaviour created an option
using each item as both value *and* label: which worked fine if the list
was intended to look like this:

    $options = array('Apple', 'Banana', 'Coconut');
    <option value="Apple">Apple</option>
    <option value="Banana">Banana</option>
    <option value="Coconut">Coconut</option>

However, this didn't suit associative arrays very well, which would've
been passed in like this:

    $options = array(
        'AU' => 'Australia',
        'NZ' => 'New Zealand',
        'DE' => 'Germany',
        'AT' => 'Austria'
    );

    <option value="Australia">Australia</option>
    <option value="New Zealand">New Zealand</option>
    ... etc

To accommodate this, a new property can be passed to SelectField's $args
hash to specify how lists of option strings are to be interpreted:

    OPTIONS_BY_VALUE: Previous behaviour. Use each string as both the
                      option's label *and* its visible label.
    OPTIONS_BY_KEY:   Will use the key each option-string is assigned as
                      its value, and the string itself as the option's
                      label. Suited for associative arrays.

Because I don't want authors to have to specify this behaviour each time
I set the default behaviour to an "intelligent guess" based on the array
keys, which should be adequate for most use-cases.